updating template check
